refactor account pages
