+++ src/app/account/register/page.tsx
@@ -1,34 +1,10 @@
 import { Metadata } from 'next';
-import Link from 'next/link';
-import { RegisterForm } from '@/app/account/register/register-form';
+import { Register } from '@/app/account/register/register';
 
 export const metadata: Metadata = {
     title: 'Register',
 };
 
 export default async function RegisterPage() {
-    return (
-        <main className="mx-auto max-w-screen-sm space-y-6 p-6">
-            <h1 className="scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-5xl">
-                Register
-            </h1>
-
-            <p className="leading-7">
-                Register for a new account to access the full range of features
-                available on this site.
-            </p>
-
-            <RegisterForm />
-
-            <p className="text-muted-foreground text-sm">
-                Already have an account?{' '}
-                <Link
-                    href="/account/login"
-                    className="underline underline-offset-4"
-                >
-                    Log in now
-                </Link>
-            </p>
-        </main>
-    );
+    return <Register />;
 }
