+++ src/app/account/register/register-form.tsx
@@ -1,17 +1,12 @@
 'use client';
 
-import { useRouter } from 'next/navigation';
-import { useAction } from 'next-safe-action/hooks';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { z } from 'zod';
-import { toast } from 'sonner';
-import { registerAction } from '@/app/account/actions/register-action';
 import { Form } from '@/components/ui/form';
 import { LoadingButton } from '@/components/loading-button';
 import { DateFormField } from '@/components/date-form-field';
 import { TextFormField } from '@/components/text-form-field';
-import { ServerActionErrorMessage } from '@/components/server-action-error';
 import { isInPast } from '@/lib/dates';
 
 const schema = z
@@ -46,11 +41,14 @@ const schema = z
         path: ['confirmPassword'],
     });
 
-type RegisterFormValues = z.infer<typeof schema>;
+export type RegisterFormValues = z.infer<typeof schema>;
 
-export function RegisterForm() {
-    const router = useRouter();
+type RegisterFormProps = {
+    loading?: boolean;
+    onSubmit: (data: RegisterFormValues) => void;
+};
 
+export function RegisterForm({ loading, onSubmit }: RegisterFormProps) {
     const form = useForm<RegisterFormValues>({
         resolver: zodResolver(schema),
         defaultValues: {
@@ -62,20 +60,6 @@ export function RegisterForm() {
             dateOfBirth: '',
         },
     });
-
-    const { execute, isPending } = useAction(registerAction, {
-        onSuccess: () => {
-            toast.success('User successfully registered.');
-            router.push('/account/login');
-        },
-        onError: ({ error }) => {
-            toast.error(<ServerActionErrorMessage result={error} />);
-        },
-    });
-
-    function onSubmit(data: RegisterFormValues) {
-        execute(data);
-    }
 
     return (
         <Form {...form}>
@@ -92,7 +76,7 @@ export function RegisterForm() {
                     maxLength={254}
                     autoComplete="username"
                     required
-                    disabled={isPending}
+                    disabled={loading}
                 />
 
                 <div className="flex flex-col gap-6 sm:flex-row">
@@ -104,7 +88,7 @@ export function RegisterForm() {
                         maxLength={50}
                         autoComplete="new-password"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                     <TextFormField
@@ -114,7 +98,7 @@ export function RegisterForm() {
                         maxLength={50}
                         autoComplete="new-password"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                 </div>
@@ -127,7 +111,7 @@ export function RegisterForm() {
                         maxLength={50}
                         autoComplete="given-name"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                     <TextFormField
@@ -136,7 +120,7 @@ export function RegisterForm() {
                         maxLength={50}
                         autoComplete="family-name"
                         required
-                        disabled={isPending}
+                        disabled={loading}
                         className="flex-1"
                     />
                 </div>
@@ -147,11 +131,11 @@ export function RegisterForm() {
                     label="Date Of Birth"
                     autoComplete={['bday-day', 'bday-month', 'bday-year']}
                     required
-                    disabled={isPending}
+                    disabled={loading}
                 />
 
                 <div className="flex flex-col-reverse justify-end gap-2 sm:flex-row">
-                    <LoadingButton type="submit" loading={isPending}>
+                    <LoadingButton type="submit" loading={loading}>
                         Submit
                     </LoadingButton>
                 </div>
